// src/types/materialConsumptionItem.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialConsumptionItemDTO {
	id: number;
	line_num: number;
	material_consumption_id: number;
	material_id: number;
	measure_unit_id: number;
	quant: number;
}

export interface MaterialConsumptionItem {
	id: number;
	line_num: number;
	material_consumption_id: number;
	material_id: number;
	measure_unit_id: number;
	quant: number;
}

export type MaterialConsumptionItemKey = Pick<
	MaterialConsumptionItem,
	"id"
>;
export type MaterialConsumptionItemNew = Pick<
	MaterialConsumptionItem,
	"line_num" | "material_consumption_id" | "material_id" | "measure_unit_id" | "quant"
>;
export type MaterialConsumptionItemUpd = Partial<Pick<
	MaterialConsumptionItem,
	"line_num" | "material_consumption_id" | "material_id" | "measure_unit_id" | "quant"
>>;

export interface MaterialConsumptionItemUpdate {
	key: MaterialConsumptionItemKey;
	model: MaterialConsumptionItemUpd;
}