// src/types/materialReceiptItem.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialReceiptItemDTO {
	id: number;
	line_num: number;
	material_receipt_id: number;
	material_id: number;
	measure_unit_id: number;
	construction_site_id: number | null;
	quant: number;
	price: number;
	amount: number;
	vat_percent: number;
	vat_amount: number;
}

export interface MaterialReceiptItem {
	id: number;
	line_num: number;
	material_receipt_id: number;
	material_id: number;
	measure_unit_id: number;
	construction_site_id: number | null;
	quant: number;
	price: number;
	amount: number;
	vat_percent: number;
	vat_amount: number;
}

export type MaterialReceiptItemKey = Pick<
	MaterialReceiptItem,
	"id"
>;
export type MaterialReceiptItemNew = Pick<
	MaterialReceiptItem,
	"line_num" | "material_receipt_id" | "material_id" | "measure_unit_id" | "construction_site_id" | "quant" | "price" | "amount" | "vat_percent" | "vat_amount"
>;
export type MaterialReceiptItemUpd = Partial<Pick<
	MaterialReceiptItem,
	"line_num" | "material_receipt_id" | "material_id" | "measure_unit_id" | "construction_site_id" | "quant" | "price" | "amount" | "vat_percent" | "vat_amount"
>>;

export interface MaterialReceiptItemUpdate {
	key: MaterialReceiptItemKey;
	model: MaterialReceiptItemUpd;
}