// src/types/materialTransferItem.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialTransferItemDTO {
	id: number;
	line_num: number;
	material_transfer_id: number;
	material_id: number;
	measure_unit_id: number;
	quant: number;
}

export interface MaterialTransferItem {
	id: number;
	line_num: number;
	material_transfer_id: number;
	material_id: number;
	measure_unit_id: number;
	quant: number;
}

export type MaterialTransferItemKey = Pick<
	MaterialTransferItem,
	"id"
>;
export type MaterialTransferItemNew = Pick<
	MaterialTransferItem,
	"line_num" | "material_transfer_id" | "material_id" | "measure_unit_id" | "quant"
>;
export type MaterialTransferItemUpd = Partial<Pick<
	MaterialTransferItem,
	"line_num" | "material_transfer_id" | "material_id" | "measure_unit_id" | "quant"
>>;

export interface MaterialTransferItemUpdate {
	key: MaterialTransferItemKey;
	model: MaterialTransferItemUpd;
}